mgr/cephadm: add placementspec for which hosts get ceph.conf
Add a config option to control which hosts (by default, *) get a
ceph.conf (if the bool manage_etc_ceph_ceph_conf option is enabled).

We don't modify the existing option because changing a type makes for a
messy migration: we have to sort out which section the config option is
in to change it.  Also, a simple on/off which is more friendly than
specifying "*" to enable something.

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit f1096030d9b6caeeaf7c1982b770b5f3fbe913bc)