osd/osd_type: use f->dump_unsigned() when appropriate
it is more explicit than `dump_stream()`, as we can not tell if dumped
variable is an integer or not by reading the code. it helps us to figure
out the scheme of the dumped object.

Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 22128e3de697f3fdf66faf3fe3b701a3a599968f)