doc/cephadm: improve "Canceling an Upgrade"
This PR improves the section "Canceling an Upgrade"
in the "Upgrading Ceph" chapter of the cephadm
documentation.

I removed an extraneous prompt and rewrote a sentence
so that it was congruent with other sentences in similar
places elsewhere in the documentation.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit ee455f00ab26240e47d7e8588615f386c05ed271)