doc/cephadm: rewrite troubleshooting 1 of x
This PR improves the readability and format
of the troubleshooting.rst file. This also
makes a change to the markdown of one of the
sub-subsections so that it is made of tildes
(~) instead of carets (^), because that's
the RST standard.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit bd24b6d392f4c652d8721eccb14c12197d617ca6)