test: fix wrong alarm
This is not a serious bug. Please look over the following scenario
in HitSetWrite test:

1. append object, which causes to add the object to the hitset
2. call hit_set_get to get the object accessed recently
3. see the hitset to check that appened object is in the hitset

After step 2, assert occurs because plpg_on_pool_change,
which invokes hit_set_clear, is called
between step 1 and step 2.
So, the object this unit test want to add to the hitset
is not in the hitset sometime.
To avoid this, this commit adds retry logic.

Fixes: https://tracker.ceph.com/issues/45423

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>
(cherry picked from commit 9a020694d43628a16e31f2e051088da5d98553e8)