mgr/cephadm: enable prometheus module before deploying prometheus
The mon will restart the mgr when the module is enabled, so we don't
really have to do anything here.  The raise is there just in case the
mgr doesn't immediately get the new mgrmap and respawn, although there is
likely no harm done if we continue to deploy prometheus in the meantime,
even if we're interrupted partway through.

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit 50dc1d0decb2fbf7b9129bddba940969410be5cd)